cleaned up history type categories
Records in HISTORY with undefined history categories were identified using checks.R.
Three records in history used a histtype code that is not used in the database; these were fixed.

Also, changed the convention that "*" at the end of histtype name for regrowth or establishment events indicates that regrowth and establishment were assumed to occur in the same year. This is now indicated in the est.regrowth.assumed.same.year field in HISTORY. Metadata and Matlab code were updated accordingly.