added notes on Lei-T-111 values
Lei-T-111 ANPP_woody_branch is suspect in that it is an unusually high outlier. I checked it against Luyssaert, and it has been correctly transposed. Original soruce is NRA. Did not remove value, but noted that it is suspect in the notes. It is a managed site that will be excluded from analyses of natural forests.