fixed Las Margaritas biomass error
This error was detected by Isabel Cano:

Last weeks we have been comparing the output of ESMs with ForC data and wanted to check with you about a location in Mexico with a very high AGB value that our model highlights as an outlier. The entry corresponds to:

measurement.ID = 8049

sites.sitename = “Las Margaritas”

citation.ID = “Poorter_2015_decs”

The original reference is Poorter et al. 2015 GEB  (https://onlinelibrary.wiley.com/doi/abs/10.1111/geb.12364) and I understand the data was extracted from a supplementary table given in their Appendix S1. The value is 853.0 (AGB, Mg/ha), that corresponds to the value of 0.47 * 853.0 = 400.91 Mg of C/ha in ForC simplified. Plot size is not specified but the length and width (50 and 2! m) do not match the expected 0.1 ha. My guess is that plot size was instead 50 x 20 = 1000 m^2 and as a consequence the AGB figure given in the table was overestimated by a factor of 10. The typo seems to affect only that table since there is no such high AGB in their Figure 2 in Southern Mexico.